timer name update and controller tweak for better compatibility
